ARROW-13735: [C++][Python] Creating a Map array with non-default field names segfaults
The segfault only happened in Debug build (which is why some weren't able to repro) because it was from a failing DCHECK on this line:

https://github.com/apache/arrow/blob/e734856676a00335f3dfe79899a995d87286f5a9/cpp/src/arrow/array/array_nested.cc#L193

The LHS of the DCHECK is missing field names, while the RHS has them.

The basic problem is that MapBuilder doesn't preserve field names when creating the array. This PR fixes that.

Closes #11855 from wjones127/ARROW-13735-map-custom-name-segfault

Lead-authored-by: Will Jones <willjones127@gmail.com>
Co-authored-by: Will Jones <wjones127@users.noreply.github.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>
(cherry picked from commit 31a07be1d9dc2f7c9720cc0fdcd7f083d947aba1)